+++ src/Expression/Name.php
@@ -11,16 +11,25 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/twig-front-matter
  */
 
-namespace Ergebnis\Twig\FrontMatter;
+namespace Ergebnis\Twig\FrontMatter\Expression;
 
-final class Example
+use Ergebnis\Twig\FrontMatter\Exception;
+
+final class Name
 {
     private function __construct(private readonly string $value)
     {
     }
 
+    /**
+     * @throws Exception\InvalidName
+     */
     public static function fromString(string $value): self
     {
+        if (1 !== \preg_match('/^[a-z][0-9a-z_]*$/i', $value)) {
+            throw Exception\InvalidName::invalid($value);
+        }
+
         return new self($value);
     }