+++ CHANGELOG.md
@@ -8,4 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`7039e81...main`][7039e81...main].
 
+### Added
+
+- Implemented `Expression\Name` as a value object ([#1]), by [@localheinz]
+
 [7039e81...main]: https://github.com/ergebnis/twig-front-matter/compare/7039e81...main
+
+[#1]: https://github.com/ergebnis/twig-front-matter/pull/1
+
+[@localheinz]: https://github.com/localheinz